Improve verification of dropped items, and make all plants drop themselves too
Currently, plants that are configured to drop anything else than
themselves (e.g. grass which drops wheat seeds), never drop themselves, even
if they don't drop the 'other' item. This was fixed.

Also, items that don't exist are no longer dropped

Instead of checking for the farming mod, all dropped nodes are
verified for existence.

A shorthand was added for including the default dropping
in a plant's drop-list